Fix MariaDB startup
Fix MariaDB startup by skipping grant table to ensure root the
ability to log in, and not randomizing the root password via
mysqladmin as it seems like this now turns off root access via socket
auth.

The initial --skip-grant-tables should be safe since networking is
disabled and the socket is only available in the local /tmp
directory.